move also applyModifiedVersion to this new comp
