+++ scopes/component/checkout/checkout-version.ts
@@ -1,21 +1,14 @@
-import * as path from 'path';
 import type { Consumer } from '@teambit/legacy.consumer';
 import type { ComponentID } from '@teambit/component-id';
 import type { Version } from '@teambit/objects';
 import type { SourceFile } from '@teambit/component.sources';
 import { RemovePath, DataToPersist } from '@teambit/component.sources';
-import type { PathOsBased } from '@teambit/legacy.utils';
 import { pathNormalizeToLinux } from '@teambit/legacy.utils';
 import type { ConsumerComponent } from '@teambit/legacy.consumer-component';
 import { BitError } from '@teambit/bit-error';
 import chalk from 'chalk';
-import type {
-  ApplyVersionResult,
-  FilesStatus,
-  MergeStrategy,
-  MergeResultsThreeWay,
-} from '@teambit/component.modules.merge-helper';
-import { FileStatus, MergeOptions } from '@teambit/component.modules.merge-helper';
+import type { ApplyVersionResult, FilesStatus, MergeResultsThreeWay } from '@teambit/component.modules.merge-helper';
+import { applyModifiedVersion, FileStatus, MergeOptions } from '@teambit/component.modules.merge-helper';
 import type { CheckoutProps } from './checkout.main.runtime';
 
 export type ComponentStatusBase = {
@@ -144,87 +137,6 @@ export async function removeFilesIfNeeded(
     }
   });
   await dataToPersist.persistAllToFS();
-}
-
-/**
- * relevant only when
- * 1) there is no conflict => add files from mergeResults: addFiles, overrideFiles and modifiedFiles.output.
- * 2) there is conflict and mergeStrategy is manual => add files from mergeResults: addFiles, overrideFiles and modifiedFiles.conflict.
- *
- * this function only updates the files content, it doesn't write the files
- */
-export function applyModifiedVersion(
-  componentFiles: SourceFile[],
-  mergeResults: MergeResultsThreeWay,
-  mergeStrategy: MergeStrategy | null | undefined
-): { filesStatus: Record<string, any>; modifiedFiles: SourceFile[] } {
-  let modifiedFiles = componentFiles.map((file) => file.clone());
-  const filesStatus = {};
-  if (mergeResults.hasConflicts && mergeStrategy !== MergeOptions.manual) {
-    return { filesStatus, modifiedFiles };
-  }
-  mergeResults.modifiedFiles.forEach((file) => {
-    const filePath: PathOsBased = path.normalize(file.filePath);
-    const foundFile = modifiedFiles.find((componentFile) => componentFile.relative === filePath);
-    if (!foundFile) throw new BitError(`file ${filePath} not found`);
-    if (file.conflict) {
-      foundFile.contents = Buffer.from(file.conflict);
-      filesStatus[file.filePath] = FileStatus.manual;
-    } else if (typeof file.output === 'string') {
-      foundFile.contents = Buffer.from(file.output);
-      filesStatus[file.filePath] = FileStatus.merged;
-    } else if (file.isBinaryConflict) {
-      // leave the file as is and notify the user later about it.
-      foundFile.contents = file.fsFile.contents;
-      filesStatus[file.filePath] = FileStatus.binaryConflict;
-    } else {
-      throw new BitError(`file ${filePath} does not have output nor conflict`);
-    }
-  });
-
-  mergeResults.addFiles.forEach((file) => {
-    const filePath: PathOsBased = path.normalize(file.filePath);
-    if (modifiedFiles.find((m) => m.relative === filePath)) return;
-    modifiedFiles.push(file.fsFile);
-    filesStatus[file.filePath] = FileStatus.added;
-  });
-  mergeResults.deletedConflictFiles.forEach((file) => {
-    if (!file.fsFile) return;
-    const filePath: PathOsBased = path.normalize(file.filePath);
-    if (modifiedFiles.find((m) => m.relative === filePath)) return;
-    modifiedFiles.push(file.fsFile);
-    filesStatus[file.filePath] = FileStatus.added;
-  });
-  mergeResults.removeFiles.forEach((file) => {
-    const filePath: PathOsBased = path.normalize(file.filePath);
-    filesStatus[file.filePath] = FileStatus.removed;
-    modifiedFiles = modifiedFiles.filter((f) => f.relative !== filePath);
-  });
-  mergeResults.remainDeletedFiles.forEach((file) => {
-    const filePath: PathOsBased = path.normalize(file.filePath);
-    modifiedFiles = modifiedFiles.filter((f) => f.relative !== filePath);
-    filesStatus[file.filePath] = FileStatus.remainDeleted;
-  });
-  mergeResults.deletedConflictFiles.forEach((file) => {
-    filesStatus[file.filePath] = FileStatus.deletedConflict;
-  });
-
-  mergeResults.overrideFiles.forEach((file) => {
-    const filePath: PathOsBased = path.normalize(file.filePath);
-    const foundFile = modifiedFiles.find((componentFile) => componentFile.relative === filePath);
-    if (!foundFile) throw new BitError(`file ${filePath} not found`);
-    foundFile.contents = file.fsFile.contents;
-    filesStatus[file.filePath] = FileStatus.overridden;
-  });
-  mergeResults.updatedFiles.forEach((file) => {
-    const filePath: PathOsBased = path.normalize(file.filePath);
-    const foundFile = modifiedFiles.find((componentFile) => componentFile.relative === filePath);
-    if (!foundFile) throw new BitError(`file ${filePath} not found`);
-    foundFile.contents = file.content;
-    filesStatus[file.filePath] = FileStatus.updated;
-  });
-
-  return { filesStatus, modifiedFiles };
 }
 
 export function throwForFailures(allComponentsStatus: ComponentStatusBase[]) {
