+++ scopes/component/checkout/index.ts
@@ -4,13 +4,12 @@ export type { CheckoutMain, CheckoutProps } from './checkout.main.runtime';
 export default CheckoutAspect;
 export { CheckoutAspect };
 
-export {
-  applyModifiedVersion,
-  applyVersion,
-  removeFilesIfNeeded,
-  updateFileStatus,
-  throwForFailures,
-} from './checkout-version';
+export { applyVersion, removeFilesIfNeeded, updateFileStatus, throwForFailures } from './checkout-version';
+
+// backward compatibility
+// export {
+//   applyModifiedVersion
+// } from '@teambit/component.modules.merge-helper';
 
 export type { ComponentStatus, ComponentStatusBase, ApplyVersionWithComps } from './checkout-version';
 export { checkoutOutput } from './checkout-cmd';
