+++ scopes/component/modules/merge-helper/index.ts
@@ -29,3 +29,5 @@ export {
   getAddedOutput,
   getWorkspaceConfigUpdateOutput,
 } from './merge-output';
+
+export { applyModifiedVersion } from './apply-version';
